feat: 도메인 설계

+++ peauty-persistence/src/main/java/com/peauty/persistence/payment/OrderEntity.java
@@ -21,7 +21,7 @@ public class OrderEntity {
     private Long threadId;
 
     @Column(name = "cost", nullable = false)
-    private Integer cost;
+    private Long cost;
 
     @Column(name = "payment_status", nullable = false)
     private Boolean isPaymentCompleted;
